fix: Todoモデルをドメイン層に移動
