+++ Stampy/View/Todo/TodoView.swift
@@ -7,45 +7,6 @@
 
 import SwiftUI
 
-struct Todo : Identifiable, Equatable {
-    let id: UUID
-    let user_id: UUID
-    let title: String
-    let state: TodoState
-    
-    static let Empty: Todo = .init(id: UUID(), user_id: UUID(), title: "", state: .NotYet)
-    static let ExampleYet: Todo = .init(id: UUID(), user_id: UUID(), title: "Example", state: .NotYet)
-    static let ExampleDone: Todo = .init(id: UUID(), user_id: UUID(), title: "ExampleComplete!!", state: .Done)
-    
-    enum TodoState {
-        case NotYet
-        case Done
-    }
-    
-    var isDone: Bool {
-        state == .Done
-    }
-    
-    var isEmpty: Bool {
-        return !isDone && title.isEmpty
-    }
-    
-    func newTitle(_ title: String) -> Self {
-        .init(id: id, user_id: user_id, title: title, state: state)
-    }
-}
-
-extension Todo.TodoState: CaseIterable {
-    var description: String {
-        switch self {
-        case .NotYet:
-            return "未完了"
-        case .Done:
-            return "完了"
-        }
-    }
-}
-
 struct TodoView: View, TodoDelegate {
     @ObservedObject var viewmodel: TodoViewModel = .init()
